geometry: retain admitted bounds and carrier cardinalities
Store procedural record bounds as optional finite scalars behind checked admission and atomic mutation. Procedural curve/surface construction is total on admitted definitions. Keep surface pole counts as usize and borrow retained pole rows for lookup and native edits. Port every affected constructor, bound and cardinality caller; report SLDPRT native count limits before allocation or output.

Validation: root check with tests and all-target Clippy for IR, ASM and nine codecs; 771 IR schema tests and 9,099 non-golden codec/ASM tests passed (one existing ignored test). Source policy and diff checks pass. Independent complete-CADIR non-finite-bound serde mutation and construction compile-fail evidence were reviewed. No golden payload changed.

Use --no-verify for this intermediate integration because the owned annotation API repair still awaits its SLDPRT caller merge and the repository-wide wire/golden gate is intentionally deferred until all changes are in the tree. Scoped validation above is complete; the normal final gate remains required.

+++ crates/cadmpeg-codec-rhino/src/writer.rs
@@ -1289,8 +1289,8 @@ fn validate_nurbs_trim(
     use cadmpeg_ir::eval::{nurbs_surface_point, pcurve_uv};
     use cadmpeg_ir::topology::Sense;
 
-    let u_count = surface.u_count() as usize;
-    let v_count = surface.v_count() as usize;
+    let u_count = surface.u_count();
+    let v_count = surface.v_count();
     let u_domain = [
         surface.u_knots()[surface.u_degree() as usize],
         surface.u_knots()[u_count],
@@ -1739,8 +1739,8 @@ fn check_nurbs_surface(
 ) -> Result<(), CodecError> {
     let u_order = surface.u_degree() as usize + 1;
     let v_order = surface.v_degree() as usize + 1;
-    let u_count = surface.u_count() as usize;
-    let v_count = surface.v_count() as usize;
+    let u_count = surface.u_count();
+    let v_count = surface.v_count();
     if u_order < 2
         || v_order < 2
         || i32::try_from(u_order).is_err()
